Исправления

+++ src/main/java/ru/mail/polis/homework/simple/DoubleTask.java
@@ -15,8 +15,8 @@ public class DoubleTask {
      */
     public static String equation(int a, int b, int c) {
         double discriminantSqrt = Math.sqrt(b * b - 4 * a * c);
-        double x1 = -b / (2.0 * a) + (discriminantSqrt / Math.abs(2 * a));
-        double x2 = -b / (2.0 * a) - (discriminantSqrt / Math.abs(2 * a));
+        double x1 = (double) -b / (2 * a) + (discriminantSqrt / Math.abs(2 * a));
+        double x2 = (double) -b / (2 * a) - (discriminantSqrt / Math.abs(2 * a));
         return x1 + ", " + x2;
     }
